switch Infection to Fragment
